Integrate Search into Index
